move `SelectableCapture` into `stream`
In the future the `stream.rs` file will only be compiled on non-windows
platforms. `SelectableCapture` is an implementation detail of how
stream is implemented on those platforms, so it makes sense to move
`SelectableCapture` into that module.
