+++ src/lib.rs
@@ -1281,7 +1281,7 @@ impl<T: Activated + ?Sized> Capture<T> {
         if !self.nonblock {
             return Err(NonNonBlock);
         }
-        PacketStream::new(SelectableCapture::new(self)?, codec)
+        PacketStream::new(self, codec)
     }
 
     /// Sets the filter on the capture using the given BPF program string. Internally this is
@@ -1395,31 +1395,6 @@ impl AsRawFd for Capture<Active> {
         let fd = unsafe { raw::pcap_fileno(self.handle.as_ptr()) };
         assert!(fd != -1, "Unable to get file descriptor for live capture");
         fd
-    }
-}
-
-/// Newtype [`Capture`] wrapper that exposes `pcap_get_selectable_fd()`.
-#[cfg(feature = "capture-stream")]
-struct SelectableCapture<T: State + ?Sized> {
-    inner: Capture<T>,
-    fd: RawFd,
-}
-
-#[cfg(feature = "capture-stream")]
-impl<T: Activated + ?Sized> SelectableCapture<T> {
-    fn new(capture: Capture<T>) -> Result<Self, Error> {
-        let fd = unsafe { raw::pcap_get_selectable_fd(capture.handle.as_ptr()) };
-        if fd == -1 {
-            return Err(InvalidRawFd);
-        }
-        Ok(Self { inner: capture, fd })
-    }
-}
-
-#[cfg(all(unix, feature = "capture-stream"))]
-impl<T: Activated + ?Sized> AsRawFd for SelectableCapture<T> {
-    fn as_raw_fd(&self) -> RawFd {
-        self.fd
     }
 }
